content/html/document/test/test_bug446483.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=446483
-->
<head>
  <title>Test for Bug 446483</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=446483">Mozilla Bug 446483</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 446483 **/

function gc() {
  SpecialPowers.gc();
}

function runTest() {
  document.getElementById('display').innerHTML =
    '<iframe src="bug446483-iframe.html"><\/iframe>\n' +
    '<iframe src="bug446483-iframe.html"><\/iframe>\n';

  setInterval(gc, 1000);

  setTimeout(function() {
    document.getElementById('display').innerHTML = '';
    ok(true, '');
    SimpleTest.finish();
  }, 4000);
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

</script>
</pre>
</body>
</html>
